Fix linter and unit tests in Darwin
